Split basis function classes into multiples files. Fix redundant bias check on bersntein basis. Fix docstring math notation
