+++ docs/code/basis-function.md
@@ -4,5 +4,17 @@ template: overrides/main.html
 
 # Documentation for `Basis Functions`
 
-::: sysidentpy.basis_function._basis_function
+::: sysidentpy.basis_function._bersntein
+      show_root_heading: false
+
+::: sysidentpy.basis_function._bilinear
+      show_root_heading: false
+
+::: sysidentpy.basis_function._fourier
+      show_root_heading: false
+
+::: sysidentpy.basis_function._legendre
+      show_root_heading: false
+
+::: sysidentpy.basis_function._polynomial
       show_root_heading: false
