+++ sysidentpy/basis_function/__init__.py
@@ -1,2 +1,7 @@
-from ._basis_function import Polynomial, Fourier, Bersntein
-from .legendre_basis import Legendre
+from ._legendre import Legendre
+from ._polynomial import Polynomial
+from ._fourier import Fourier
+from ._bersntein import Bersntein
+from ._bilinear import Bilinear
+
+__all__ = ["Bersntein", "Bilinear", "Fourier", "Legendre", "Polynomial"]
